requirements.txt
bert-for-tf2
tensorflow
numpy
